Not show image. Naming convention changed.
We used to just override to "predictions.jpg" for every one. In order
to allow for multiple files, we had to do two things:
1) Not attempt opening the file from the linux machine over ssh.
2) Allow different naming of the next image as you iterate, so that the
file is saved down separately.

We will use this in a script to take the raw images and convert it into
the processed images.

TODO: Only pick up "person" class.